moving tutorial parsing into checkdocs
